backend-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
torch==2.0.1
transformers==4.34.1
sentence-transformers==2.2.2
llama-cpp-python==0.2.19
numpy==1.24.3
pydantic==2.4.2
